Modified a subset of class member function signatures for classes AnansiMDState, AnansiMDStateISE, and AnansiMDStatePCL.
Modified signature of function AnansiMDState::initializeSimulationEnvironment to be a const function.
Modified signature of function AnansiMDState::processCommandLine to be a const function.
Modified signature of function AnansiMDState::_initializeSimulationEnvironment to be a const function.
Modified signature of function AnansiMDState::_processCommandLine to be a const function.
Modified signature of function AnansiMDStateISE::_initializeSimulationEnvironment to be a const function.
Modified signature of function AnansiMDStatePCL::_processCommandLine to be a const function.